test: relax JIT setup in misc.getmetrics test
To test whether <jit_snap_restore> metric is counted right for side
exits <minstitch> JIT engine parameter is tighted to prevent side trace
compilation (considering <math.fmod> being used for the corresponding
branch of execution, the side trace will try to compile with stitching).
However, if the test is run on the platform configured with the option
LUAJIT_ENABLE_CHECKHOOK enabled, the resulting trace exceeds the
hardcoded <minstitch> limit and the metric is miscounted due to this new
trace. To resolve this issue <minstitch> limit is set to the maximum
amount of IRRef in the trace, preventing trace compilation for any
LuaJIT configuration.

Relates to tarantool/tarantool#7762

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>