ci: change runner dispatch for LuaJIT testing
Before the patch, all LuaJIT testing jobs were dispatched to the runner
with the name given via the strategy matrix <RUNNER> entry. However, a
more convenient way for dispatching is the triplet of the following
labels:
* <matrix.ARCH> to specify the host architecture (i.e. x86_64 or ARM64)
* <matrix.OS> to specify the OS family name (either Linux or macOS)
* 'self-hosted' to make job dispatch only to the runners from Tarantool
  ghacts-shared-* pool and never choose GitHub public resources

With the given triplet the proper runner is chosen for the particular
testing flavor.

Moreover, the new 'lightweight' and 'regular' labels are also used in
<runs-on> list, since the new "lightweight" runners have been introduced
to ghacts-shared-* pool. There are a couple of LuaJIT tests that require
more memory than is provided by "lightweight" runners, so only "regular"
ones need to be chosen for full LuaJIT testing. At the same time there
is no need to use "regular" runner for LuaJIT Static analysis workflow,
hence 'lightweight' label is added to <runs-on> list in lint.yml.

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>