ci: use strategy matrix for integration workflow
Considering the changes made in the scope of the previous commit, it has
been decided to dispatch LuaJIT integration testing to the runner the
similar way (i.e. by using the pair of <matrix.ARCH> + <matrix.OS>
values). As a result a new job with quite similar strategy matrix[1] is
added as a successor for <test-luajit> job. The new job uses the new
version (with <inputs.arch> and <inputs.os> parameters introduced and
the optional <inputs.host> obsolete parameter) of the LuaJIT integration
workflow from the tarantool/tarantool repo.

[1]: https://docs.github.com/en/actions/using-workflows/reusing-workflows#using-a-matrix-strategy-with-a-reusable-workflow

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>