ci: introduce workflow for exotic builds
This workflow is well-suited for test specific builds like dual-number
mode on x86_x64, build with disabled JIT or FFI, etc.

For now, just the dual-number mode is added, due to the need to test
future commits. If there is a need to add a new build type, then the
`FLAVOR` column of CI matrix should be extended. Also, only Linux x86_64
is tested, because we want to decrease resource usage in CI.

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>