test: fix lua-Harness JIT-related tests
lua-Harness considers whether JIT is enabled or not in the scope of the
403-jit.t and 411-luajit.t tests. However, the original condition is
wrong, since <jit.status> yields false for both cases, when JIT is just
turned off and when LuaJIT is built without compiler support. So, if
<jit.status> yields false, the latter case is considered. The condition
is fixed to differ both aforementioned cases in the following way: when
jit.opt is nil, LuaJIT is built without compiler; otherwise, JIT support
is on board.

Part of tarantool/tarantool#8252

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>