test: add utility for parsing `jit.dump`
This commit adds a simple parser for traces to be analyzed in the test.
For now nothing too fancy at all -- just start `jit.dump` to a temporary
file and parse and remove this file when dump is stopped. The array with
resulting traces is returned.

For now, just find a single IR by pattern and return ref number and IR,
if they exist. More functionality may be added if it is necessary for
tests.

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>