Fix frame for on-trace out-of-memory error.
Reported by ruidong007.

(cherry-picked from commit 2d8300c1944f3a62c10f0829e9b7847c5a6f0482)

When an on-trace OOM error is triggered from a frame that is child in
regard to `jit_base`, and `L->base` is not updated correspondingly
(FUNCC, for example), it is possible to encounter an inconsistent Lua
stack in the error handler.

This patch adds a fixup for OOM errors on the trace that always sets the
Lua stack base to `jit_base`, so the stack is now consistent.

Maxim Kokryashkin:
* added the description and the test for the problem

Part of tarantool/tarantool#8825

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>