test: fix flaky <unit-jit-parse.test.lua>
This commit is the follow-up for
5570eb824116e124b5d3442d57a0e90011bcddd8 ("test: add utility for parsing
`jit.dump`")

This test is flaky on GC64 build due to some unpredictable JIT
compilation hotcount and trace recording.

This patch adds `jit.off()`/`jit.on()` calls to be sure that the JIT is
enabled only for the particular part of the test, where we collect the
trace dump.

Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>