test: introduce test:done TAP helper
Before the patch, all tests (except <lj-603-err-snap-restore.test.lua>)
in tarantool-tests suite terminate their execution via <os.exit> with
the status depending on the test results. However, the second argument
of <os.exit> was omitted, and Lua universe was not properly finalized as
a result. This behaviour becomes a problem, when LuaJIT is built with
LUAJIT_USE_SYSMALLOC option and AddressSanitizer support enabled, since
the sanitizer starts reporting false positive errors about the memory
allocations without the corresponding memory releases. To resolve these
errors, the second parameter to each <os.exit> call terminating the test
has to be added.

To avoid loss of the aforementioned parameter in the future, <test:done>
helper has been added to the TAP module. Depending on the single
parameter, the new helper either properly finalize the test being run,
or simply checks all the test assertions and raises an error if any of
them fail. The latter case is added especially to handle the specifics
of <lj-603-err-snap-restore.test.lua> and still check that everything
works fine.

Needed for tarantool/tarantool#5878

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>