ci: support coveralls
The patch adds an action that installs gcovr and a workflow that runs
regression test suites, produces a summary of current code coverage and
sends code coverage data to Coveralls. Coveralls is a web service that
lets you inspect every detail of your coverage. See Tarantool's LuaJIT
page on Coveralls [1].

1. https://coveralls.io/github/tarantool/luajit

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>