sysprof: improve parser's memory footprint
This patch reduces sysprof's parser memory footprint, by avoiding
reading all callchains before collapsing them. Instead of it, parser
merges stacks immediately after reading them and stores counts in a Lua
table.

The `collapse.lua` module is purged as a result of the patch, but it is
left as a stub to keep the integration testing intact. This stub should
be removed in the next series.

Resolves tarantool/tarantool#8700

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>