gdb: fix Python <assert> statement usage
This <assert> misuse was introduced in the initial implementation of GDB
extension within the commit 58790750b9c4bd4c21d883f109ab552a2e202a15
("gdb: introduce luajit-gdb extension") and also propagated to the
initial implementation of LLDB extension by copy-pasting in scope of the
commit 62fc84a8f89b8e5650162ba1c7696b0f84cf5c25 ("lldb: introduce
luajit-lldb"). However, <assert> is not a function, but statement, so
parenthesis around the condition are considered as a tuple constructor.
This patch simply removes the excess parenthesis to finally follow
Python semantics of the <assert> statement.

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>