Fix frame for more types of on-trace error messages.
Thanks to Maxim Kokryashkin.

(cherry-picked from commit d5bbf9cdb4c5eddc404a90bd44f077cfb3a57a90)

This patch fixes the same issue with the frame as the previous one, but
now for the table overflow error in the `err_msgv` function. The test
for the problem uses the table of GC finalizers, although they are not
required to reproduce the issue. They are only used to make the test as
simple as possible.

Maxim Kokryashkin:
* added the description and the test for the problem

Resolves tarantool/tarantool#562
Part of tarantool/tarantool#8825

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>