test: fix flaky <unit-jit-parse.test.lua> again
This commit is the follow-up for
8e46d60104057159812b25b91a36c029220e87f5 ("test: fix flaky
<unit-jit-parse.test.lua>")

This test is flaky on the Tarantool build due to hotcount collisions
with `test:plan()` function. This patch fixes it by moving the set of
the hotloop option right before the test trace recording to reset
hotcount right before it, so there are no possible collisions at all.

Signed-off-by: Igor Munkin <imun@tarantool.org>