test: set dependencies in BuildTestCLib macro
In test suite `tarantool-tests`, we are using the CMake macro
`BuildTestCLib` for building auxiliary test C libraries. This macro
collects the dependencies required for tarantool tests in `TESTLIBS`
list. However, if this macro is used several times in one file, only the
last library is added to the dependency list, since the `TESTLIBS` is
updated only in the parent scope.

The patch removes `TESTLIBS` list and adds each test the C library as a
dependency to CMake target `tarantool-tests` explicitly using the
`add_dependencies` command.

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>