test: add test for conversions folding
This patch adds the test for commit
1a401622fe83ae695226c94df3416e312315d659 ("Fix assertions."). This patch
removes incorrect assertions in the fold optimizations for conversions
from numbers to different integer types. Since the issue affects only
branch 2.0, there is no need to fix it. Nevertheless, the test is
required to avoid regressions in the future.

Part of tarantool/tarantool#9145

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>