test: prevent hanging Tarantool CI by sysprof test
Sysprof tests often require long running loops to be executed, so a
certain situation is likely to occur. E.g the test added in scope of the
commit 285a1b0a16c1c3224ab93f7cf1fef17ab29e4ab4 ("sysprof: fix crash
during FFUNC stream") expects FFUNC VM state at the moment when
stacktrace is being collected. Unfortunately, it leads to test routine
hang for particular environments. Hence, it was decided to disable the
test at least for Tarantool CI, until the issue is not resolved.

Relates to tarantool/tarantool#9387

Signed-off-by: Igor Munkin <imun@tarantool.org>