test: adjust lua-Harness test error assertion
The current pattern in 303-package.t for the test case loading invalid
Lua script is too tight for the following reasons revealed by Tarantool
testing:
* The platform might prepend the common prefix to all Lua errors, so
  anchoring the pattern to the beginning of the error message is too
  strict for matching the particular errors.
* The platform might canonicalize module names (i.e., use absolute
  paths) for searchers and loaders, so matching the current working
  directory via "./" is too straight-forward.

Hence, this patch removes the starting anchor from the pattern and fixes
the last part of it to match only the "basename" of the module being
loaded. As a result, the check is not too strict and Lua-specific but
rather starts following common sense.

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>