test: prepare lauxilarily libs for LuaJIT-tests
This patch adds rules to build C and C++ libs from the
<LuaJIT-tests/src/> directory. Also, it allows these libraries to be
loaded via `require()` (since `luaL_openlib()` evolves global state).
The name of "ctest" library is changed to "libctest" to allow it to be
loaded via both `ffi.load()` and `require()`. "cpptest" library is
renamed to "libcpptest" for consistency.

Part of tarantool/tarantool#9398

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>