test: enable <ffi_callback.lua> in LuaJIT-tests
This patch names all subtests and includes the test in <index>.
The test with calls with 10 arguments with sizes less than 8 bits are
dummy for M1. This suite lacks a mechanism to skip subtests satisfying
complicated conditions, so it is done manually by the corresponding `if`
check.

Part of tarantool/tarantool#9398

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>