test: enable <ffi_jit_call.lua> in LuaJIT-tests
This patch uses the basename of the ctest library to be loaded via FFI.
It adds groups with the names of subtests and enables the test in
<index>. The corresponding conditions are changed to the tags checked by
analogy with other tests. Tests with calls with 10 arguments with sizes
less than 8 bits are dummy for M1. This suite lacks a mechanism to skip
subtests satisfying complicated conditions, so it is done manually by
the corresponding `if` check.

Part of tarantool/tarantool#9398

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>