test: enable <ffi_new.lua> in LuaJIT-tests
This patch changes `dofile()` (which uses an unreliable relative file
path and mutates `_G` with global functions to be defined) to the
corresponding require of the helper function. It adds groups with the
names of subtests and enables the test in <index>. It renames structures
typedefs to avoid conflicts with cdef in other tests. The GC threshold
is increased since all tests run in a single process, so the number of
GC objects is greater.

Part of tarantool/tarantool#9398

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>