codehealth: add `nd` to the codespell ignore list
This patch is a follow-up to the
15acf5b2ff3841e757434e63394de97662a17e81 ("Fix zero stripping in %g
number formatting."). `nd` is used as a stack buffer for number
formatting and mentioned in the test comment. This patch adds it to the
ignore list to avoid false-positives for comments like "nd ==> and, 2nd"
for codespell version 2.2.6.

Also, it sorts the ignore list for better readability.

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>