test: fix initialization in lj-549-lua-load.test.c
This patch is the follow up for the test introduced in the commit
cf2d9bb89121c5a22846e71cc21c67f383d74b7b ("Followup fix for embedded
bytecode loader.") to make old C compilers happy.

In C language, objects with static storage duration have to be
initialized with constant expressions or with aggregate initializers
containing constant expressions. The term "constant" refers to literal
constants (like 1, 'a', 0xFF, and so on), enum members, and the results
of such operators as sizeof. Const-qualified objects (of any type) are
not constants in C language terminology. They cannot be used in
initializers of objects with the static storage duration, regardless of
their type.

Hence, the patch defines the first one byte from the Lua bytecode
signature explicitly and prevents the source code discrepancy with the
corresponding static assertion.

Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>