test: introduce routine to build error message
LuaJIT modules yield the plain errors, but LuaJIT binary adds 'luajit:'
prefix to it, so 411-luajit.t test in lua-Harness suite expects the
error message with the aforementioned prefix in the corresponding
assertions. At the same time, Tarantool doesn't prepend errors produced
by the LuaJIT module with anything.

To tweak the pattern to be used within the 411-luajit.t chunk, the
auxiliary error building function is introduced in this patch.

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>